requirements.txt
py-tools-ds==0.24.1
arosics==1.13.2
geoarray==0.19.2

numpy
pandas
scipy
scikit-learn
joblib
matplotlib
tqdm
pyyaml

rasterio
rioxarray
xarray
shapely
pyproj
netCDF4
spectral
pystac-client
earthaccess
POT
git+https://github.com/EnSpec/hytools.git

torch>=2.0
wandb
einops
timm
